Move fixtures into dedicated test files
