+++ python/tests/test_fabric.py
@@ -2,7 +2,15 @@ from unittest.mock import Mock
 import pytest
 import stix2
 
+import vast
 from vast.utils.stix import IDENTITIES
+
+
+@pytest.fixture
+def fabric():
+    config = vast.utils.config.create()
+    backbone = vast.fabric.backbones.InMemory()
+    return vast.fabric.Fabric(config, backbone)
 
 
 @pytest.mark.asyncio
